Add missing documentation files mafft*.1
The files mafft-homologs.1 and mafft.1 (that are documentation files)
are required by the Makefile.
The Makefile simply copies these files to the installation directory.
If the files are missing the make is unsuccessful.
The files are not automatically generated, but are given in the
sources (even if the extension .1 makes you think otherwise)

The files are missing from the github repo LARA2-mafft.
I have taken the files from the original sources at:
https://mafft.cbrc.jp/alignment/software/mafft-7.453-with-extensions-src.tgz